requirements.txt
# Core dependencies for WaveGAN and DCGAN
torch>=1.9.0
torchaudio>=0.9.0
numpy>=1.20.0

# For metrics collection and analysis
pandas>=1.3.0
matplotlib>=3.3.0

# For image processing (spectrograms)
Pillow>=8.0.0

# For system monitoring (optional but recommended)
psutil>=5.8.0

# For data processing
scipy>=1.7.0

# Progress bars
tqdm>=4.60.0
